+++ src/resources/js/Components/app/PostItem.vue
@@ -1,4 +1,5 @@
 <script setup>
+import { Disclosure, DisclosureButton, DisclosurePanel } from '@headlessui/vue';
 
     defineProps({
         post: Object
@@ -21,7 +22,19 @@
                 <small class="text-gray-400">{{ post.created_at }}</small>
             </div>
         </div>
-        <div v-html="post.body">
+        <div>
+            <Disclosure v-slot="{ open }">
+                <div v-if="!open" v-html="post.body.substring(0, 200)" />
+                <DisclosurePanel>
+                    <div v-html="post.body" />
+                </DisclosurePanel>
+                <div class="flex justify-end">
+                    <DisclosureButton class="text-blue-500">
+                        {{ open ? 'Read less' : 'Read more' }}
+                    </DisclosureButton>
+                </div>
+            </Disclosure>
+        
         </div>
     </div>
 </template>